Initial upload
This is my entire home website at 31.12.2025 - based on CC4.0 licence which allows you to develop this with or without me if you take it as a science, it's more complicated if it's business and might ask you to buy their work from people. Here you find pdf's and links to my other apps and pages at this stage.